Renaming RDF files
When going with releases, the versions will be encoded in the release
and should not be contained in the source files. Therefore, the
previously used naming format with date and version number is not
useful and may confuse the reader. The updated version represents the
nature of the files as the currently valid truth, while versions and
releases will be documented separately using the respective Github
Release functionality.