Add back checking on commit messages
This patch re-introduced the checks on commit messages in order to make
the git log more uniform and thus easier to read and follow.

Originally, we turned off the check of commit messages in
2ec5d79b71de83be9fb364c39df66a5cb5a61569. Many people were confused with
the uninformative error messages whenever the style check did not pass.
We hope that this change provides a better error description and finds
more frequent adoption.

We also rename the existing CI workflow to clarify the distinction.