Create Initial Commit for draft rc02 schemas (#112)
With this patch, we start specifying RC02 schemas. This is an initial 
commit. 

All the schemas and examples are taken from **RC01**. This was
necessary because we wanted to have the continuous integration running.
The following commits are expected to bluntly overwrite both 
the schemas and the examples.

In this patch, we wanted to:

* Get the continuous integration running,
* Warn the readers in the README that the schemas and examples are
  for review-only, and
* Clean up the examples and documentation so that only the minimal 
  examples remain.

Unfortunately, we had to recreate the pull request because there was 
confusion about what the scope of the pull request was and caused too
many inconsistent changes in #100.