Add missing constraints in JSON (#148)
We add a couple of missing constraints in JSON schema according to the
book. Notably, we restrict string length to at least 1 character and add
various patterns for strings.

Additionally, we update some of the examples where minor fixes were
possible, and remove the examples where the number of errors was too
high.