Merge feature/RdfUpdate and master
We make a check out of the RDF and SHACL schemas from the branch
`feature/RdfUpdate` which underwent multiple iterations of reviews
(against the book, against the JSON schema and against the generated
schema based on aas-core-codegen).

Since the branch `feature/RdfUpdate` diverged too much from the master,
we do not bother to apply Git magic, but check out the files by "brute
force".