Merge examples from `feature/RdfUpdate` and master (#165)
The branch `feature/RdfUpdate` diverged too much from the master so we
had to manually merge it into the master in the pull request #158.
Unfortunately, we forgot to merge in the examples by mistake.

In this patch, we check out the examples from `feature/RdfUpdate` and
merge them into the master. We could not bother to do it with "git
magic" so we used a "brute-force" approach.

This is related to the issue #164.