Sort definitions in JSON schema (#171)
We sort definitions in JSON schema to make it easier for future diffs
against, say, V3RC02. Also, RDF and SHACL schemas are sorted, so we
follow the convention with JSON as well. Finally, it is easier to find a
definition in a sorted list of definitions by hand.

We checked that the schema order only changed by:

```
cat aas.new.json|sort > /tmp/a
cat aas.json|sort > /tmp/b
diff /tmp/a /tmp/b
```