Address comments in JSON readme from #220
We adress comments and make the respective changes in the JSON readme
raised in the review of the pull request #220.

The section about JSON representation of `ValueDataType` is left
unfixed, as it needs to be properly specified. For now, values are *not*
going to be represented as JSON strings, but as JSON boolean and number
in cases where applicable.