Update Readmes in schema examples
The readmes were outdated and referred to the previous examples which
were manually written. We automatically generate the examples now, so we
update the readmes accordingly.

At this point, to make the readmes consistent, we remove the manually
written example for the RDF schema. If we eventually want to restore it,
please do not forget to update the corresponding readme.

Fixes #241.