Remove unneccessary `modelType`-requirements from JSON-Schema
The `modelType` discriminator is needed for distinction of
otherwise indistinguishible objects.
Previously, we required the `modelType`-discriminator
in places, where it actually wasn't needed, namely in
abstract classes, that are never instantiated.

We improved the aas-core-codegen JSON generation algorithm
so that these occurrances are removed.

These objects previously had a `modelType`-discriminator
and have them now removed:

- `DataElement`
- `SubmodelElement`
- `EventElement`
- `Identifiable`
- `RelationshipElement_abstract`
- `HasDataSpecification`

Note, that the concrete children of these objects do still
have the `modelType`-discriminator, if they need them.