Schema Remove Pattern for PathType
Currently, the regex for Path type is
non-compliant to the specification v3.0,
as it does not allow for AASX packages
to be written.

As a temporary bugfix in v3.0.1, it was decided
to remove the invariant check, as changing the
pattern would result in a breaking change.
This is of course no final solution,
there will be a better fix for version 3.1.

See #299