Define classes in separate PUMLs
Before classes
were defined multiple times in different
PUML files. The problem is that, if some
attribute of a class changes, we need to
adapt all the files where the class was
defined. Now we define classes
in separate files und use !include command
in PUMLs where the class will be used.