Integrate PlantUML diagrams + small fixes (#463)
* Add PUML diagrams

* Integrate PUMLs into ADOCs

* Fix PUML with packages

* Add idta puml theme

* Define classes in separate PUMLs

Before classes
were defined multiple times in different
PUML files. The problem is that, if some
attribute of a class changes, we need to
adapt all the files where the class was
defined. Now we define classes
in separate files und use !include command
in PUMLs where the class will be used.

* Add and integrate PUMLs in `uml.adoc`

* Fix figure refs

* Add labeled-association and Reference repr

* Fix Reference associations

* Assign attributes to parent classes

* Replace relative paths with absolute