Remove ":page-partial:" from .adocs
- :page-partial: was set in multiple .adoc files,
which are not partials, but standalone docs.

When you set the :page-partial: attribute,
it usually means that the current file
is not meant to be a standalone document,
but rather a reusable component that will
be included elsewhere.

- Include constraints after the doc name.
If smth is included before doc name, it blocks
automatic reuse of the doc name in xrefs
where the doc is referenced